backend/requirements.txt
fastapi==0.135.1
pydantic==2.12.5
requests==2.32.5
pandas==3.0.1
openpyxl==3.1.5
python-multipart==0.0.22
google-cloud-secret-manager==2.26.0
websockets==16.0
slowapi==0.1.9
google-cloud-bigquery==3.40.1
python-jose==3.5.0
uvicorn==0.42.0
cloud-sql-python-connector==1.20.1
sqlalchemy==2.1.0b1
pg8000==1.31.5
google-api-python-client==2.193.0
google-auth==2.49.1
gspread==6.2.1
apscheduler==3.11.2
Pillow==12.2.0
google-cloud-storage==3.10.1
google-cloud-pubsub==2.37.0
alembic==1.18.4